requirements-updated.txt
# Talk World 3.0 - Python Dependencies
# Updated requirements with correct Coqui TTS version and PyTorch compatibility

# Core TTS and STT
TTS==0.22.0
openai-whisper>=20231117

# PyTorch (compatible with TTS 0.22.0)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and NLP
transformers>=4.20.0
sentencepiece>=0.1.99
tokenizers>=0.13.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.10.0

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Translation models
torch-audio>=0.13.0
sacremoses>=0.0.53
fairseq>=0.12.2

# Utilities
requests>=2.28.0
tqdm>=4.64.0
Pillow>=9.0.0